// src/features/availability/domain/enums.ts
export enum AvailabilityStatus {
  ONLINE = 'ONLINE',
  OFFLINE = 'OFFLINE',
  BUSY = 'BUSY',
}

export enum SessionMode {
  MANUAL = 'MANUAL',
  AUTO = 'AUTO',
}

export enum SessionType {
  CHAT = 'CHAT',
  CALL = 'CALL',
}

export enum WorkingDay {
  MONDAY = 'MONDAY',
  TUESDAY = 'TUESDAY',
  WEDNESDAY = 'WEDNESDAY',
  THURSDAY = 'THURSDAY',
  FRIDAY = 'FRIDAY',
  SATURDAY = 'SATURDAY',
  SUNDAY = 'SUNDAY',
}